Improved regularization so that it accepts arbitrary linear operators.
