+++ setup.py
@@ -86,7 +86,8 @@ setup(name='scikit-fda',
                         'scikit-datasets[cran]>=0.1.24',
                         'rdata',
                         'cython',
-                        'mpldatacursor'],
+                        'mpldatacursor',
+                        'multimethod>=1.2'],
       setup_requires=pytest_runner,
       tests_require=['pytest'],
       test_suite='tests',
