+++ skfda/misc/__init__.py
@@ -1,4 +1,4 @@
 from . import covariances, kernels, metrics
+from . import operators
 from . import regularization
-from ._lfd import LinearDifferentialOperator
 from ._math import log, log2, log10, exp, sqrt, cumsum, inner_product
