+++ skfda/misc/operators/__init__.py
@@ -1,0 +1,2 @@
+from ._linear_differential_operator import LinearDifferentialOperator
+from ._operators import Operator, gramian_matrix, gramian_matrix_optimization
