+++ skfda/misc/regularization/__init__.py
@@ -1,4 +1,2 @@
-from ._endpoints_difference_regularization import EndpointsDifferenceRegularization
-from ._l2_regularization import L2Regularization
-from ._linear_diff_op_regularization import LinearDifferentialOperatorRegularization
-from ._regularization import Regularization, compute_penalty_matrix
+from ._regularization import (TikhonovRegularization,
+                              compute_penalty_matrix)
